% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.cf_performance}
\alias{coef.cf_performance}
\title{Coefficient Method for cf_performance Objects}
\usage{
\method{coef}{cf_performance}(object, ...)
}
\arguments{
\item{object}{A \code{cf_performance} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
The point estimate.
}
\description{
Coefficient Method for cf_performance Objects
}
