% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_methods.R
\name{coef.tr_performance}
\alias{coef.tr_performance}
\title{Coefficient Method for tr_performance Objects}
\usage{
\method{coef}{tr_performance}(object, ...)
}
\arguments{
\item{object}{A \code{tr_performance} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
The point estimate.
}
\description{
Coefficient Method for tr_performance Objects
}
