% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_sens_spec.R
\name{.bootstrap_cf_sens_spec_crossfit}
\alias{.bootstrap_cf_sens_spec_crossfit}
\title{Bootstrap for cross-fitted sens/spec}
\usage{
.bootstrap_cf_sens_spec_crossfit(
  predictions,
  outcomes,
  treatment,
  covariates,
  threshold,
  treatment_level,
  metric,
  n_folds,
  n_boot,
  conf_level,
  propensity_learner = NULL,
  outcome_learner = NULL,
  parallel,
  ncores,
  ps_trim_spec = NULL
)
}
\description{
Bootstrap for cross-fitted sens/spec
}
\keyword{internal}
