% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_learner.R
\name{.fit_ml_learner}
\alias{.fit_ml_learner}
\title{Fit an ML Learner to Data}
\usage{
.fit_ml_learner(learner, formula, data, family = "binomial")
}
\arguments{
\item{learner}{An \code{ml_learner} object.}

\item{formula}{Model formula.}

\item{data}{Training data.}

\item{family}{Character: "binomial" for classification, "gaussian" for
regression.}
}
\value{
A fitted model object with class \code{ml_fitted}.
}
\description{
Fit an ML Learner to Data
}
\keyword{internal}
