% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_learner.R
\name{.predict_ml_learner}
\alias{.predict_ml_learner}
\title{Predict from Fitted ML Learner}
\usage{
.predict_ml_learner(object, newdata, type = "response", ...)
}
\arguments{
\item{object}{An \code{ml_fitted} object.}

\item{newdata}{Data frame for prediction.}

\item{type}{Character: "response" for probabilities (default), "class" for
class predictions.}

\item{...}{Ignored.}
}
\value{
Numeric vector of predicted probabilities.
}
\description{
Predict from Fitted ML Learner
}
\keyword{internal}
