% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{.stratified_folds}
\alias{.stratified_folds}
\title{Create Stratified Folds}
\usage{
.stratified_folds(y, K)
}
\arguments{
\item{y}{Binary outcome vector}

\item{K}{Number of folds}
}
\value{
Vector of fold assignments
}
\description{
Create Stratified Folds
}
\keyword{internal}
