% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{.summarize_cv_results}
\alias{.summarize_cv_results}
\title{Summarize CV Results}
\usage{
.summarize_cv_results(results_df, metric)
}
\arguments{
\item{results_df}{Data frame of fold-level results}

\item{metric}{Performance metric}
}
\value{
Data frame with summary statistics
}
\description{
Summarize CV Results
}
\keyword{internal}
