% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_learner.R
\name{is_ml_fitted}
\alias{is_ml_fitted}
\title{Check if Object is a Fitted ML Model}
\usage{
is_ml_fitted(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical indicating if \code{x} is an \code{ml_fitted} object.
}
\description{
Check if Object is a Fitted ML Model
}
\keyword{internal}
