% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_learner.R
\name{is_ml_learner}
\alias{is_ml_learner}
\title{Check if Object is an ML Learner Specification}
\usage{
is_ml_learner(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical indicating if \code{x} is an \code{ml_learner} object.
}
\description{
Check if Object is an ML Learner Specification
}
\keyword{internal}
