% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cf_auc}
\alias{plot.cf_auc}
\title{Plot Method for cf_auc Objects}
\usage{
\method{plot}{cf_auc}(x, ...)
}
\arguments{
\item{x}{A \code{cf_auc} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A plot object.
}
\description{
Creates an ROC curve plot.
}
