% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{plot.tr_roc}
\alias{plot.tr_roc}
\alias{plot.cf_roc}
\title{Plot ROC Curve}
\usage{
\method{plot}{tr_roc}(
  x,
  add_diagonal = TRUE,
  show_naive = TRUE,
  main = NULL,
  col = "blue",
  naive_col = "gray50",
  lwd = 2,
  ...
)

\method{plot}{cf_roc}(
  x,
  add_diagonal = TRUE,
  show_naive = TRUE,
  main = NULL,
  col = "blue",
  naive_col = "gray50",
  lwd = 2,
  ...
)
}
\arguments{
\item{x}{An object of class \code{tr_roc} or \code{cf_roc}.}

\item{add_diagonal}{Logical indicating whether to add a diagonal reference
line (representing random classifier). Default is TRUE.}

\item{show_naive}{Logical indicating whether to show the naive ROC curve
if available. Default is TRUE.}

\item{main}{Title for the plot.}

\item{col}{Color for the ROC curve. Default is "blue".}

\item{naive_col}{Color for the naive ROC curve. Default is "gray50".}

\item{lwd}{Line width. Default is 2.}

\item{...}{Additional arguments passed to plot().}
}
\value{
Invisibly returns the input object.
}
\description{
Creates a plot of the ROC curve.
}
\examples{
set.seed(123)
n <- 500
x <- rnorm(n)
a <- rbinom(n, 1, plogis(-0.5 + 0.5 * x))
y <- rbinom(n, 1, plogis(-1 + x - 0.5 * a))
pred <- plogis(-1 + 0.8 * x)

roc <- cf_roc(
  predictions = pred,
  outcomes = y,
  treatment = a,
  covariates = data.frame(x = x),
  n_thresholds = 51
)
plot(roc)
}
