% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{print.cf_compare}
\alias{print.cf_compare}
\title{Print Method for cf_compare Objects}
\usage{
\method{print}{cf_compare}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{cf_compare} object.}

\item{digits}{Number of digits to print.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for cf_compare Objects
}
