% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{print.cf_cv}
\alias{print.cf_cv}
\title{Print Method for cf_cv Objects}
\usage{
\method{print}{cf_cv}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{cf_cv} object.}

\item{digits}{Number of digits to print.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for cf_cv Objects
}
