% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance.R
\name{print.cf_nuisance}
\alias{print.cf_nuisance}
\title{Print Method for cf_nuisance Objects}
\usage{
\method{print}{cf_nuisance}(x, ...)
}
\arguments{
\item{x}{A \code{cf_nuisance} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for cf_nuisance Objects
}
