% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_learner.R
\name{print.ml_learner}
\alias{print.ml_learner}
\title{Print Method for ml_learner Objects}
\usage{
\method{print}{ml_learner}(x, ...)
}
\arguments{
\item{x}{An \code{ml_learner} object.}

\item{...}{Ignored.}
}
\description{
Print Method for ml_learner Objects
}
